% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.topo.normalised.R
\name{dist.topo.normalised}
\alias{dist.topo.normalised}
\title{Normalised topological distance between two trees}
\usage{
dist.topo.normalised(tr1, tr2, nrand = 100)
}
\arguments{
\item{tr1}{A rooted phylogenetic tree of class \code{"phylo"}.}

\item{tr2}{A rooted phylogenetic tree of class \code{"phylo"} with the same
tip labels as \code{tr1}.}

\item{nrand}{Integer. Number of random permutations used to estimate the
maximum distance. Default \code{100}.}
}
\value{
A numeric scalar in \code{[0, 1]}: the normalised topological
  distance between the two trees.
}
\description{
Computes the topological distance between \code{tr1} and \code{tr2} (using
\code{\link[ape]{dist.topo}}) and normalises it by the maximum distance
observed over \code{nrand} random permutations of \code{tr2}'s tip labels.
}
\examples{
library(ape)
set.seed(1)
tr1 <- rcoal(10)
tr2 <- rcoal(10)
tr2$tip.label <- tr1$tip.label
dist.topo.normalised(tr1, tr2, nrand = 20)

}
\seealso{
\code{\link[ape]{dist.topo}}, \code{\link{distunlab}}
}
