% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.monophyletic.R
\name{find.monophyletic}
\alias{find.monophyletic}
\title{Find monophyletic clades matching a tip-label pattern}
\usage{
find.monophyletic(tr, tag, include.singletons = FALSE)
}
\arguments{
\item{tr}{A rooted phylogenetic tree of class \code{"phylo"}.}

\item{tag}{Character string (regular expression) to match against tip
labels.}

\item{include.singletons}{Logical. If \code{TRUE}, tips that match
\code{tag} but do not belong to any monophyletic multi-tip clade are also
included in the output. Default \code{FALSE}.}
}
\value{
A list of character vectors. Each element is a vector of tip labels
  belonging to one monophyletic clade. If \code{include.singletons = TRUE},
  singleton matching tips are appended as individual character scalars.
}
\description{
Identifies all maximal monophyletic clades whose tip labels all match a
given pattern. The algorithm traverses the tree from the root downward,
pruning subtrees that have no matching tips, and collecting subtrees where
every tip matches.
}
\examples{
library(ape)
set.seed(1)
tr <- rtree(12)
tr$tip.label[c(1, 2, 5)] <- paste0("TAG_", tr$tip.label[c(1, 2, 5)])
find.monophyletic(tr, "TAG_")
find.monophyletic(tr, "TAG_", include.singletons = TRUE)

}
\seealso{
\code{\link{get.mrca}}, \code{\link{is.polytomy}}
}
