% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.deepest.br.length.R
\name{get.deepest.br.length}
\alias{get.deepest.br.length}
\title{Total length of root-descending branches}
\usage{
get.deepest.br.length(tr)
}
\arguments{
\item{tr}{A rooted phylogenetic tree of class \code{"phylo"}.}
}
\value{
A numeric scalar: the sum of root-descending branch lengths.
}
\description{
Returns the sum of the lengths of all branches that descend directly from
the root node.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(10)
get.deepest.br.length(tr)

}
\seealso{
\code{\link{get.oldest.branch.length}}
}
