% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.df.R
\name{get.df}
\alias{get.df}
\title{Branch-length imbalance statistic}
\usage{
get.df(tre)
}
\arguments{
\item{tre}{A phylogenetic tree of class \code{"phylo"}.}
}
\value{
A numeric scalar.
}
\description{
Computes an imbalance statistic based on mean-centred branch lengths: the
sum of centred external branch lengths minus the sum of all centred branch
lengths, which simplifies to the negative sum of centred internal branch
lengths.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(10)
get.df(tr)

}
