% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.first.node.height.R
\name{get.first.node.height}
\alias{get.first.node.height}
\title{Age of the third-oldest internal node}
\usage{
get.first.node.height(tr)
}
\arguments{
\item{tr}{A rooted phylogenetic tree of class \code{"phylo"} with branch
lengths in units of time.}
}
\value{
A named numeric scalar: the age of the third-oldest internal node.
}
\description{
Returns the age of the third-oldest internal node (i.e. the third entry
when internal-node times are sorted in decreasing order). Used as a
summary statistic for tree shape.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(10)
get.first.node.height(tr)

}
\seealso{
\code{\link{intnode.times}}
}
