% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mrca.R
\name{get.mrca}
\alias{get.mrca}
\title{Most recent common ancestor of a set of tips}
\usage{
get.mrca(tr, tips)
}
\arguments{
\item{tr}{A rooted phylogenetic tree of class \code{"phylo"}.}

\item{tips}{Integer vector of tip indices (\code{1..Ntips}).}
}
\value{
A single integer: the node index of the MRCA, following ape's
  convention (\code{Ntips+1..Ntips+Nnodes} for internal nodes).
}
\description{
Finds the most recent common ancestor (MRCA) node of a given set of tips.
When a single tip is supplied the function returns its direct parent node.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(10)
get.mrca(tr, c(1L, 3L, 5L))

}
\seealso{
\code{\link{get.ancestor.nodes.branches}}, \code{\link{find.sister}}
}
