% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ordinates.R
\name{get.ordinates}
\alias{get.ordinates}
\title{Node plotting coordinates for a phylogenetic tree}
\usage{
get.ordinates(tr)
}
\arguments{
\item{tr}{A rooted phylogenetic tree of class \code{"phylo"}.}
}
\value{
A numeric matrix with columns \code{x.coord} (node age),
  \code{node.index} (ape node index), and \code{y.coord} (plotting
  ordinate).
}
\description{
Computes x and y plotting coordinates for every node (tips and internal
nodes) of a phylogenetic tree. x coordinates are node ages (from
\code{\link{all.node.times}}); y coordinates are based on the average
ordinate of direct descendants, assigned from leaves upward. Used
internally by \code{\link{plot.tree.lines}}.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(10)
head(get.ordinates(tr))

}
\seealso{
\code{\link{plot.tree.lines}}, \code{\link{all.node.times}}
}
