% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.rtt.dist.R
\name{get.rtt.dist}
\alias{get.rtt.dist}
\title{Root-to-tip distances for all tips}
\usage{
get.rtt.dist(phylo)
}
\arguments{
\item{phylo}{A rooted phylogenetic tree of class \code{"phylo"} with branch
lengths in units of time (or substitutions per site).}
}
\value{
A named numeric vector of length \code{Ntips} with root-to-tip
  distances. Names are the integer tip indices (\code{1..Ntips}).
}
\description{
Returns the patristic distance from the root to each tip. For ultrametric
trees all values are equal; for non-ultrametric (heterochronous) trees they
vary.
}
\examples{
library(ape)
set.seed(1)
tr <- rtree(10)
get.rtt.dist(tr)

}
\seealso{
\code{\link{node.to.tip.dist}}, \code{\link{allnode.times}}
}
