% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intnode.times.R
\name{intnode.times}
\alias{intnode.times}
\title{Ages of internal nodes in a phylogenetic tree}
\usage{
intnode.times(phylo)
}
\arguments{
\item{phylo}{A rooted phylogenetic tree of class \code{"phylo"} with branch
lengths in units of time.}
}
\value{
A named numeric vector of length \code{Nnodes}. Names are the
  integer internal-node indices (\code{Ntips+1..Ntips+Nnodes}).
}
\description{
Returns the ages (divergence times) of internal nodes only, expressed as
time before the most recent tip. Equivalent to the forward-time distances
from root used internally by other NELSI functions.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(10)
intnode.times(tr)

}
\seealso{
\code{\link{allnode.times}}
}
