% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.polytomy.R
\name{is.polytomy}
\alias{is.polytomy}
\title{Test whether a node is a polytomy}
\usage{
is.polytomy(tr, node)
}
\arguments{
\item{tr}{A phylogenetic tree of class \code{"phylo"}.}

\item{node}{Integer. The index of the node to test, using ape's convention:
tips are \code{1..Ntips}, internal nodes are \code{Ntips+1..Ntips+Nnodes}.}
}
\value{
A logical scalar.
}
\description{
Returns \code{TRUE} if the given node has more than two direct descendants
(i.e. is a polytomy / multifurcation), and \code{FALSE} if it is a
bifurcation or a tip.
}
\examples{
library(ape)
tr <- read.tree(text = '((a, b, c), (d, (e, f)));')
is.polytomy(tr, 8)   # the (a, b, c) node -- TRUE
is.polytomy(tr, 9)   # the root -- FALSE

}
\seealso{
\code{\link{find.monophyletic}}
}
