% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.lsd.dates.R
\name{make.lsd.dates}
\alias{make.lsd.dates}
\title{Write a sampling-date file for LSD}
\usage{
make.lsd.dates(tr, grep.sep = ".+_", outfile = "out.date", random = F)
}
\arguments{
\item{tr}{A phylogenetic tree of class \code{"phylo"} whose tip labels
encode sampling dates (e.g. \code{"taxon_2015.5"}).}

\item{grep.sep}{Character. Regular expression matched and removed from each
tip label to obtain the date string. Default \code{".+_"} strips
everything up to and including the last underscore.}

\item{outfile}{Character. Path to the output file. Default \code{"out.date"}.}

\item{random}{Logical. If \code{TRUE}, dates are randomly shuffled across
tips before writing. Default \code{FALSE}.}
}
\value{
\code{NULL} invisibly. The function is called for its side effect
  of writing \code{outfile}.
}
\description{
Extracts sampling dates from tip labels and writes them to a tab-delimited
file in the format expected by LSD (Least Squares Dating). Dates are
extracted by stripping the portion of each tip label matched by
\code{grep.sep}.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(5)
tr$tip.label <- paste0("taxon_", c(2010, 2011, 2012, 2013, 2014))
make.lsd.dates(tr, outfile = tempfile())

}
