% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.edge.ages.R
\name{mid.edge.ages}
\alias{mid.edge.ages}
\title{Midpoint ages of all branches}
\usage{
mid.edge.ages(phylo)
}
\arguments{
\item{phylo}{A rooted phylogenetic tree of class \code{"phylo"} with branch
lengths in units of time.}
}
\value{
A numeric vector of midpoint ages, one per branch, in the same
  order as \code{phylo$edge}.
}
\description{
Computes the age at the midpoint of each branch in a phylogenetic tree.
For ultrametric trees the midpoint is the daughter-node age plus half the
branch length. For non-ultrametric trees the ages are first rescaled so
that the root age equals the maximum \code{allnode.times} value, then the
same formula is applied.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(5)
mid.edge.ages(tr)

}
\seealso{
\code{\link{get.tree.data.matrix}}, \code{\link{allnode.times}}
}
