% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.to.tip.dist.R
\name{node.to.tip.dist}
\alias{node.to.tip.dist}
\title{Root-to-tip distances for all tips}
\usage{
node.to.tip.dist(phylo)
}
\arguments{
\item{phylo}{A rooted phylogenetic tree of class \code{"phylo"} with branch
lengths in units of time (or substitutions per site).}
}
\value{
A named numeric vector of length \code{Ntips} with root-to-tip
  distances. Names are the integer tip indices (\code{1..Ntips}).
}
\description{
Returns the patristic distance from the root to each tip. Equivalent to
\code{\link{get.rtt.dist}}.
}
\examples{
library(ape)
set.seed(1)
tr <- rtree(10)
node.to.tip.dist(tr)

}
\seealso{
\code{\link{get.rtt.dist}}, \code{\link{all.node.times}}
}
