% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmy.R
\name{stemmy}
\alias{stemmy}
\title{Stemminess of a phylogenetic tree}
\usage{
stemmy(tre)
}
\arguments{
\item{tre}{A phylogenetic tree of class \code{"phylo"}.}
}
\value{
A numeric scalar between 0 and 1.
}
\description{
Computes stemminess as the proportion of total branch length attributable
to internal (non-terminal) branches. For ultrametric trees all branches
leading to internal nodes are counted as internal. For non-ultrametric
trees, branches leading to tips at time 0 (contemporary tips) are treated
as external.
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(10)
stemmy(tr)

}
\seealso{
\code{\link{allnode.times}}
}
