% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.descending.branches.R
\name{sum.descending.branches}
\alias{sum.descending.branches}
\title{Total branch length of a subtree}
\usage{
\method{sum}{descending.branches}(tr, target_node)
}
\arguments{
\item{tr}{A rooted phylogenetic tree of class \code{"phylo"}.}

\item{target_node}{Integer node index (using ape's convention:
\code{1..Ntips} for tips, \code{Ntips+1..Ntips+Nnodes} for internal
nodes). Passing a tip index returns 0.}
}
\value{
A single numeric value: the total branch length below
  \code{target_node}.
}
\description{
Returns the sum of all branch lengths within the subtree rooted at
\code{target_node}, i.e. every branch below that node (excluding the branch
leading into it from above).
}
\examples{
library(ape)
set.seed(1)
tr <- rcoal(10)
# sum branches below root
sum.descending.branches(tr, length(tr$tip.label) + 1L)

}
\seealso{
\code{\link{get.descending.nodes.branches}}
}
