% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trann2trdat.R
\name{trann2trdat}
\alias{trann2trdat}
\title{Convert an annotated tree to a branch-data matrix}
\usage{
trann2trdat(tree)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{"phylo"} with a
\code{$annotations} list containing \code{length} and \code{rate_median}
elements for each branch.}
}
\value{
A numeric matrix with columns \code{branch}, \code{parent},
  \code{daughter}, \code{midage}, \code{rate}, \code{blensubs}, and
  \code{blentime}.
}
\description{
Converts a BEAST-annotated phylogenetic tree (as read by
\code{\link{read.annotated.nexus}} or \code{\link{read.annotated.tree}})
to a numeric matrix in the same format as the \code{tree.data.matrix}
element of a \code{"ratesim"} object.
}
\examples{
\dontrun{
tr <- read.annotated.nexus("beast_output.trees")
dat <- trann2trdat(tr)
}

}
\seealso{
\code{\link{read.annotated.nexus}}, \code{\link{read.annotated.tree}}
}
