% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.sister.R
\name{find.sister}
\alias{find.sister}
\title{Find tips in the sister clade}
\usage{
find.sister(tr, clade, allow.polytomy = T)
}
\arguments{
\item{tr}{A rooted phylogenetic tree of class \code{"phylo"}.}

\item{clade}{Integer vector of tip indices whose sister group is sought.}

\item{allow.polytomy}{Logical. If \code{TRUE} (default), when the MRCA is
a polytomy whose children are all in \code{clade} the function goes up
one level to find the sister.}
}
\value{
An integer vector of tip indices constituting the sister clade.
}
\description{
Identifies the tip indices in the sister group of a given set of tips.
The sister is defined as all tips descending from the parent of the MRCA
of \code{clade} that are not in \code{clade}.
}
\examples{
library(ape)
tr <- read.tree(text = '((((a, b, c, d), e), f), (g, h));')
find.sister(tr, 1)

}
\seealso{
\code{\link{get.mrca}}, \code{\link{get.descending.nodes.branches}}
}
