\name{stemmy}
\alias{stemmy}
\title{
stemmy
}
\description{
stemmy calculates the stemminess of a phylogenetic tree. 
}
\usage{
stemmy(tre)
}
\arguments{
  \item{tre}{
An object of class phylo.
}
}
\details{
None.
}
\value{
numeric.
}

\author{
David Duchene
}



## The function is currently defined as

\keyword{ molecular-clock }
